fix: moved kafkaconfigerror and validate to shared errors.py file
