+++ docling_jobkit/cli/multiproc.py
@@ -20,7 +20,7 @@ from rich.progress import (
 from docling.datamodel.service.options import ConvertDocumentsOptions
 
 from docling_jobkit.connectors.auth_context import allow_interactive_auth
-from docling_jobkit.connectors.kafka.helper import (
+from docling_jobkit.connectors.errors import (
     KafkaConfigError,
     validate_kafka_kind_pairing,
 )
